refactor: abstracted LINE, Spotify, and Cloudinary functions into individual modules
