+++ configs/cloudinary.config.js
@@ -1,0 +1,13 @@
+require('dotenv').config({path: '../.env'});    
+const cloudinaryConfig = require('cloudinary');
+
+// Hosting for Audio files
+cloudinary.config({
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY,
+    api_secret: process.env.CLOUDINARY_API_SECRET
+});
+
+module.exports = {
+    cloudinaryConfig,
+}
